+++ ChangeLog.md
@@ -1,5 +1,23 @@
 # Change Log
 
+### 2022-03-08 - 5.14.0
+
+| Library                                                                 | Min Version |
+| ----------------------------------------------------------------------- | ----------- |
+| `@autorest/core`                                                        | `3.6.2`     |
+| `@autorest/modelerfour`                                                 | `4.19.1`    |
+| `azure-core` dep of generated code                                      | `1.20.1`    |
+| `msrest` dep of generated code                                          | `0.6.21`    |
+| `azure-mgmt-core` dep of generated code (If generating mgmt plane code) | `1.3.0`     |
+
+**New Features**
+
+- Add flag `--package-mode=mgmtplane|dataplane|<custom package template folder>` to generate necessary files for package #1154
+
+**Bug Fixes**
+
+- Improve operation group documentation to prevent users from initializing operation groups themselves #1179
+
 ### 2022-03-03 - 5.13.0
 
 | Library                                                                 | Min Version |